Finally, a fully functional version.
I did a lot of refactoring to make this happen.  I am really questioning
the need for the Reading class.  It doesn't add much over a simple dict.
It isn't currently being used in the add/remove reading stuff.  I think
that it is time for some cleanup now that it is working.