Another WIP commit of refactoring.
This time I have moved the application code into readit.flaskapp.
The goal is for this module to have all of the Flask specific stuff
in it.

I also took the liberty of following the excellent recommendation
from the Flask docs with regards to configuration.  The defaults
are members of the readit.flaskapp.Application instance.  These are
overridden first by environment variables then by an optional config
file.  The defaults really should be associated with the Application
class instead of the instance... maybe that is a good next change.