Merge from a rewrite branch.
This is the culimation of a TDD rewrite.  I spent a lot of time really
learning JavaScript this time around as well as diving very deeply into
Python TDD and unit testing in general.  There were too many changes to
list easily and I've been using mercurial for my day-to-day so tracking
is really best done there.  If heroku supported mercurial as well as git,
then there would be no need whatsoever for a git repo.