Fixed requirements.txt
Apparently referencing module==dev is a bad idea in requirements.txt and I
really need the latest-and-greatest for both Flask and Flask-OpenID.  The
workaround is to pull them directly from version control so I replaced the
appropriate dependency lines with pulls from github.  Not the best solution
but it does work in a pinch.