Rewrote persistence layer yet again.
This time I moved the object IDs into the storable objects as a well-known
string property.  This cleaned up a lot of the strangeness in the Mongo
layer since we don't have to guess the object ID or synthesize them any longer.
The groundwork for removing StorableItem as a class is well underway.  The
protocol will continue to exist but the goofy implementation class will go
away completely.  Most (if not all) usage of this class has been removed.

I do need to make a pass through the documentation to make sure that it still
makes sense.  All of the unit and doc tests pass so it shouldn't be that far
off.