Added a missing ``rel="external"'' attribute to the details div.
jqTouch requires this for links that leave the sandbox.

Moved to using the newer Flask extension import model.

Previous versions of Flask required that extensions were placed into
a *namespace package* called ``flaskext''.  In a recent version (0.8 or
0.9?), a new package named ``flask.ext'' was added with some importer
magic that maps an import of ``flask.ext.X'' to ``flask_X''.  The only
affected import was the Open ID extension.