*Merged changeset 36:6bd7b7d9c3aa*
The JSON representation should have an `id` attribute.

For some reason I failed to noticed that the `object_id` property of the
Python objects were not being translating into the `id` JSON property.
The JS code was expecting the `id` property and the Python serialization
stuff was writing out the `object_id`.  Not sure how I missed this but
I noticed it when I went to implement reading deletions tonight.  It is
fixed now and properly tested across the board.